+++ fetching-service/src/main/java/rs/edu/raf/banka1/services/ListingStockService.java
@@ -29,11 +29,4 @@ public interface ListingStockService {
 
     //    returns how many new listings were added
     int addAllListingsToHistory(List<ListingHistory> listingHistories);
-
-    Optional<ListingStock> findByTicker(String ticker);
-    Optional<ListingStock> findById(Long id);
-
-    List<ListingHistory> getListingHistoriesByTimestamp(String ticker, Integer from, Integer to);
-
-    List<ListingHistory> getListingHistoriesByTimestamp(Long id, Integer from, Integer to);
 }
